review feedback & small improvements
