+++ haystack/preview/telemetry/__init__.py
@@ -1,1 +1,1 @@
-from haystack.preview.telemetry.telemetry import send_pipeline_run_event, tutorial_running, HAYSTACK_TELEMETRY_ENABLED
+from haystack.preview.telemetry._telemetry import pipeline_running, tutorial_running
