+++ haystack/preview/telemetry/_telemetry.py
@@ -1,0 +1,164 @@
+import os
+from typing import Any, Dict, Optional, TYPE_CHECKING, List, Tuple
+from pathlib import Path
+import logging
+import uuid
+import datetime
+import yaml
+import posthog
+
+from haystack.preview.telemetry._environment import dynamic_system_specs, static_system_specs
+
+if TYPE_CHECKING:
+    from haystack.preview.pipeline import Pipeline
+
+
+HAYSTACK_TELEMETRY_ENABLED = "HAYSTACK_TELEMETRY_ENABLED"
+CONFIG_PATH = Path("~/.haystack/config.yaml").expanduser()
+PIPELINE_RUN_BUFFER_SIZE = 10
+
+
+logger = logging.getLogger(__name__)
+
+
+class Telemetry:
+    """
+    Haystack reports anonymous usage statistics to support continuous software improvements for all its users.
+
+    You can opt-out of sharing usage statistics by manually setting the environment
+    variable `HAYSTACK_TELEMETRY_ENABLED` as described for different operating systems on the
+    [documentation page](https://docs.haystack.deepset.ai/docs/telemetry#how-can-i-opt-out).
+
+    Check out the documentation for more details: [Telemetry](https://docs.haystack.deepset.ai/docs/telemetry).
+    """
+
+    def __init__(self):
+        """
+        Initializes the telemetry. Loads the user_id from the config file,
+        or creates a new id and saves it if the file is not found.
+
+        It also collects system information which cannot change across the lifecycle
+        of the process (for example `is_containerized()`).
+        """
+
+        # disable posthog logging
+        for module_name in ["posthog", "backoff"]:
+            logging.getLogger(module_name).setLevel(logging.CRITICAL)
+            # Prevent module from sending errors to stderr when an exception is encountered during an emit() call
+            logging.getLogger(module_name).addHandler(logging.NullHandler())
+            logging.getLogger(module_name).propagate = False
+
+        self.user_id = None
+
+        if CONFIG_PATH.exists():
+            # Load the config file
+            try:
+                with open(CONFIG_PATH, "r", encoding="utf-8") as config_file:
+                    config = yaml.safe_load(config_file)
+                    if "user_id" in config:
+                        self.user_id = config["user_id"]
+            except Exception as e:
+                logger.debug("Telemetry could not read the config file %s", CONFIG_PATH, exc_info=e)
+        else:
+            # Create the config file
+            logger.info(
+                "Haystack sends anonymous usage data to understand the actual usage and steer dev efforts "
+                "towards features that are most meaningful to users. You can opt-out at anytime by manually "
+                "setting the environment variable HAYSTACK_TELEMETRY_ENABLED as described for different "
+                "operating systems in the [documentation page](https://docs.haystack.deepset.ai/docs/telemetry#how-can-i-opt-out). "
+                "More information at [Telemetry](https://docs.haystack.deepset.ai/docs/telemetry)."
+            )
+            CONFIG_PATH.parents[0].mkdir(parents=True, exist_ok=True)
+            self.user_id = str(uuid.uuid4())
+            try:
+                with open(CONFIG_PATH, "w") as outfile:
+                    yaml.dump({"user_id": self.user_id}, outfile, default_flow_style=False)
+            except Exception as e:
+                logger.debug("Telemetry could not write config file to %s", CONFIG_PATH, exc_info=e)
+
+        self.event_properties = static_system_specs()
+
+    def send_event(self, event_name: str, event_properties: Optional[Dict[str, Any]] = None):
+        """
+        Sends a telemetry event.
+
+        :param event_name: The name of the event to show in PostHog.
+        :param event_properties: Additional event metadata. These are merged with the
+            system metadata collected in __init__, so take care not to overwrite them.
+        """
+        event_properties = event_properties or {}
+        dynamic_specs = dynamic_system_specs()
+        try:
+            posthog.capture(
+                distinct_id=self.user_id,
+                event=event_name,
+                properties={**self.event_properties, **dynamic_specs, **event_properties},
+            )
+        except Exception as e:
+            logger.debug("Telemetry couldn't make a POST request to PostHog.", exc_info=e)
+
+
+def send_telemetry(func):
+    """
+    Decorator that sends the output of the wrapped function to PostHog.
+    The wrapped function is actually called only if telemetry is enabled.
+    """
+
+    # FIXME? Somehow, functools.wraps makes `telemetry` out of scope. Let's take care of it later.
+    def send_telemetry_wrapper(*args, **kwargs):
+        try:
+            if telemetry:
+                output = func(*args, **kwargs)
+                if output:
+                    telemetry.send_event(*output)
+        except Exception as e:
+            # Never let telemetry break things
+            logger.debug("There was an issue sending a telemetry event", exc_info=e)
+
+    return send_telemetry_wrapper
+
+
+@send_telemetry
+def pipeline_running(pipeline: "Pipeline") -> Tuple[str, Dict[str, Any]]:
+    """
+    Collects name, type and the content of the _telemetry_data attribute, if present, for each component in the
+    pipeline and sends such data to Posthog.
+
+    :param pipeline: the pipeline that is running.
+    """
+    pipeline._telemetry_runs += 1
+    if pipeline._telemetry_runs == 1 or pipeline._telemetry_runs % PIPELINE_RUN_BUFFER_SIZE == 0:
+        # Collect info about components
+        pipeline_description = pipeline.to_dict()
+        components: Dict[str, List[Dict[str, Any]]] = {}
+        for component_name, component in pipeline_description["components"].items():
+            if not component["type"] in components:
+                components[component["type"]] = []
+
+            telemetry_data = getattr(pipeline.get_component(component_name), "_telemetry_data", {})
+            if not isinstance(telemetry_data, dict):
+                telemetry_data = {"error": "_telemetry_data is present but does not contain a dict"}
+
+            components[component["type"]].append({"name": component_name, **telemetry_data})
+
+        # Data sent to Posthog
+        return "Pipeline run (2.x)", {
+            "pipeline_id": str(id(pipeline)),
+            "runs": pipeline._telemetry_runs,
+            "components": components,
+            "_unix_stamp": (datetime.datetime.now() - datetime.datetime(1970, 1, 1)).total_seconds(),
+        }
+
+
+@send_telemetry
+def tutorial_running(tutorial_id: str) -> Tuple[str, Dict[str, Any]]:
+    """
+    Send a telemetry event for a tutorial, if telemetry is enabled.
+    :param tutorial_id: identifier of the tutorial
+    """
+    return "Tutorial", {"tutorial.id": tutorial_id}
+
+
+telemetry = None
+if os.getenv("HAYSTACK_TELEMETRY_ENABLED", "true").lower() in ("true", "1"):
+    telemetry = Telemetry()
