--- haystack/preview/telemetry/environment.py
@@ -1,66 +1,0 @@
-# pylint: disable=global-statement
-import logging
-import os
-from typing import Optional
-
-from haystack import __version__
-
-logger = logging.getLogger(__name__)
-
-# IS_DOCKER cache copy
-IS_DOCKER_CACHE = None
-
-
-def set_pytorch_secure_model_loading(flag_val="1"):
-    # To load secure only model pytorch requires value of
-    # TORCH_FORCE_WEIGHTS_ONLY_LOAD to be ["1", "y", "yes", "true"]
-    os_flag_val = os.getenv("TORCH_FORCE_WEIGHTS_ONLY_LOAD")
-    if os_flag_val is None:
-        os.environ["TORCH_FORCE_WEIGHTS_ONLY_LOAD"] = flag_val
-    else:
-        logger.info("TORCH_FORCE_WEIGHTS_ONLY_LOAD is already set to %s, Haystack will use the same.", os_flag_val)
-
-
-def in_podman() -> bool:
-    """
-    Podman run would create the file /run/.containernv, see:
-    https://github.com/containers/podman/blob/main/docs/source/markdown/podman-run.1.md.in#L31
-    """
-    return os.path.exists("/run/.containerenv")
-
-
-def has_dockerenv() -> bool:
-    """
-    This might not work anymore at some point (even if it's been a while now), see:
-    https://github.com/moby/moby/issues/18355#issuecomment-220484748
-    """
-    return os.path.exists("/.dockerenv")
-
-
-def has_docker_cgroup_v1() -> bool:
-    """
-    This only works with cgroups v1
-    """
-    path = "/proc/self/cgroup"  # 'self' should be always symlinked to the actual PID
-    return os.path.isfile(path) and any("docker" in line for line in open(path))
-
-
-def has_docker_cgroup_v2() -> bool:
-    """
-    cgroups v2 version, inspired from
-    https://github.com/jenkinsci/docker-workflow-plugin/blob/master/src/main/java/org/jenkinsci/plugins/docker/workflow/client/DockerClient.java
-    """
-    path = "/proc/self/mountinfo"  # 'self' should be always symlinked to the actual PID
-    return os.path.isfile(path) and any("/docker/containers/" in line for line in open(path))
-
-
-def is_containerized() -> Optional[bool]:
-    """
-    This code is based on the popular 'is-docker' package for node.js
-    """
-    global IS_DOCKER_CACHE
-
-    if IS_DOCKER_CACHE is None:
-        IS_DOCKER_CACHE = in_podman() or has_dockerenv() or has_docker_cgroup_v1() or has_docker_cgroup_v2()
-
-    return IS_DOCKER_CACHE
